Attempt to use Cram for integration tests
Work in progress trying out two different, complementary approaches to
integration testing with Cram. The first approach basically copies the commands
already executed by the Snakefiles in the tests/builds directory into the Cram
format. The zika build, for example, is partially represented by zika.t in that
builds directory.

The second approach tries to more comprehensively test a specific augur command
with a variety of reasonable inputs. The mask.t file represents the start of
that type of test for augur mask.