Add an example test that diffs trees
Comparing trees requires comparing topology and branch lengths. Branch lengths
can be stochastically variable across tree building tools and environments for
the same input data. This commit introduces a script that will diff two Newick
trees based on the name and branch lengths of the nodes.

The diff is performed by the deepdiff package (added here as a dev dependency)
that provides an interface for limiting the significant digits considered in a
numerical comparison. To use the deepdiff tool, this script converts the
BioPython tree object into a list of nested lists with the appropriate topology.