Install augur with dev dependencies using conda's pip
conda does not provide a pip3 command, but the base Travis python does. This
means pip3 install places the augur development dependencies in the global
python environment instead of the conda environment. As a result, cram tests
fail when they rely on python packages that are only installed as dev
dependencies because cram's "python" is the conda environment's "python". This
commit should install development dependencies in the conda environment instead
of the global environment and make those packages available to the python
command inside cram tests.